Extract the LinkingLayer into a separate module
