+++ llvm-hs/llvm-hs.cabal
@@ -123,6 +123,7 @@ library
     LLVM.Internal.ObjectFile
     LLVM.Internal.OrcJIT
     LLVM.Internal.OrcJIT.CompileLayer
+    LLVM.Internal.OrcJIT.LinkingLayer
     LLVM.Internal.OrcJIT.CompileOnDemandLayer
     LLVM.Internal.OrcJIT.IRCompileLayer
     LLVM.Internal.OrcJIT.IRTransformLayer
@@ -166,6 +167,7 @@ library
     LLVM.Internal.FFI.ObjectFile
     LLVM.Internal.FFI.OrcJIT
     LLVM.Internal.FFI.OrcJIT.CompileLayer
+    LLVM.Internal.FFI.OrcJIT.LinkingLayer
     LLVM.Internal.FFI.OrcJIT.CompileOnDemandLayer
     LLVM.Internal.FFI.OrcJIT.IRCompileLayer
     LLVM.Internal.FFI.OrcJIT.IRTransformLayer
