+++ llvm-hs/src/LLVM/Internal/OrcJIT/IRTransformLayer.hs
@@ -12,7 +12,7 @@ import qualified LLVM.Internal.FFI.DataLayout as FFI
 import qualified LLVM.Internal.FFI.Module as FFI
 import qualified LLVM.Internal.FFI.OrcJIT.IRTransformLayer as FFI
 import qualified LLVM.Internal.FFI.PtrHierarchy as FFI
-import LLVM.Internal.OrcJIT hiding (ObjectLinkingLayer(..))
+import LLVM.Internal.OrcJIT
 import LLVM.Internal.OrcJIT.CompileLayer
 import LLVM.Internal.Target
